added aboutpage, style to footer
