+++ src/components/about/about.css
@@ -1,0 +1,8 @@
+.about{
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    width: 50%;
+    margin: 0 auto;
+    flex-direction: column;
+}
